requirements/testing.txt
-r base.txt

factory-boy==3.0.1
pytest==6.0.1
pytest-django==3.9.0
